1457905: Added PopulateHostedDBTask job to DELETED_JOBS
Since the PopulateHostedDBTask is not required in 2.0+ it
was deleted. After is has been run though, fetching
the job list can fail becuase it is is still in the job
history.

The job class was added to the PinsetterKernal.DELETED_JOBS
list so that instances of this job will be delted on startup