Event Simplification: Removed oldEntity from Events
- Removed the oldEntity field from the Event class, also
removed setOldEntity/getOldEntity from EventBuilder and EventFactory.
- Made setEventData on EventBuilder public to be used instead
of setOldEntity when building Events. This method will probably be
re-used similarly when newEntity will be removed in the future and
will be used when creating CREATED/DELETED events. It will be
overloaded for creating MODIFIED events (which currently need both
old/new entity).
- Removed a couple of unused methods.